refactor(sdk): consolidate retry/transport helpers, chain config, resolver constants
Move duplicated logic into the SDK so consumer packages share one source of truth:

- httpWithRetry, withRetry, delay, DEFAULT_RETRY_COUNT/DELAY_MS → @sapience/sdk/utils
  (app re-exports for unchanged import paths; api's separate withRetry with
  Infinity-retry/logger semantics stays package-local — different contract)
- POLYMARKET_RESOLVER_ADDRESSES, FORECAST_SCHEMA_UID, collectAddresses →
  @sapience/sdk/constants (replaces hardcoded EAS schema UID literal in api
  easIndexer and the locally-derived address Set in app)
- api's local etherealChain/etherealTestnetChain definitions removed in favor
  of the SDK definitions; api/services/sponsorship.ts now imports them from
  @sapience/sdk/constants. The SDK chain object adds blockExplorers metadata
  and fees.defaultPriorityFee: 1n — viem reads neither for the api's read
  paths, and the priority-fee default aligns api tx behavior with what the
  app already does.

The three viem client factories (app/api/relayer) keep their own implementations
because they have meaningfully different transport configs (multicall, retry
counts, Infura WS support) — only the shared httpWithRetry helper is hoisted.
