+++ packages/sdk/constants/index.ts
@@ -1,2 +1,3 @@
 export * from './chain';
 export * from './pythFeeds';
+export * from './resolver';
