+++ packages/sdk/utils/retry.test.ts
@@ -1,0 +1,82 @@
+import { vi, describe, it, expect, beforeEach, afterEach } from 'vitest';
+import { withRetry } from './retry';
+
+describe('withRetry', () => {
+  beforeEach(() => vi.useFakeTimers());
+  afterEach(() => vi.useRealTimers());
+
+  it('returns immediately when fn succeeds on first attempt', async () => {
+    const fn = vi.fn().mockResolvedValue('ok');
+
+    const result = await withRetry(fn);
+
+    expect(result).toBe('ok');
+    expect(fn).toHaveBeenCalledTimes(1);
+  });
+
+  it('retries and succeeds after transient failures', async () => {
+    const fn = vi
+      .fn()
+      .mockRejectedValueOnce(new Error('fail-1'))
+      .mockRejectedValueOnce(new Error('fail-2'))
+      .mockResolvedValue('ok');
+
+    const promise = withRetry(fn, 3, 100);
+
+    await vi.advanceTimersByTimeAsync(100);
+    await vi.advanceTimersByTimeAsync(200);
+
+    await expect(promise).resolves.toBe('ok');
+    expect(fn).toHaveBeenCalledTimes(3);
+  });
+
+  it('throws the last error after exhausting all retries', async () => {
+    const fn = vi
+      .fn()
+      .mockRejectedValueOnce(new Error('persistent'))
+      .mockRejectedValueOnce(new Error('persistent'))
+      .mockRejectedValueOnce(new Error('persistent'));
+
+    const promise = withRetry(fn, 2, 100);
+    const rejection = expect(promise).rejects.toThrow('persistent');
+
+    await vi.advanceTimersByTimeAsync(100);
+    await vi.advanceTimersByTimeAsync(200);
+
+    await rejection;
+    expect(fn).toHaveBeenCalledTimes(3);
+  });
+
+  it('uses exponential backoff delays', async () => {
+    const fn = vi
+      .fn()
+      .mockRejectedValueOnce(new Error('1'))
+      .mockRejectedValueOnce(new Error('2'))
+      .mockRejectedValueOnce(new Error('3'))
+      .mockResolvedValue('ok');
+
+    const promise = withRetry(fn, 3, 100);
+
+    await vi.advanceTimersByTimeAsync(99);
+    expect(fn).toHaveBeenCalledTimes(1);
+
+    await vi.advanceTimersByTimeAsync(1);
+    expect(fn).toHaveBeenCalledTimes(2);
+
+    await vi.advanceTimersByTimeAsync(200);
+    expect(fn).toHaveBeenCalledTimes(3);
+
+    await vi.advanceTimersByTimeAsync(400);
+    expect(fn).toHaveBeenCalledTimes(4);
+
+    await expect(promise).resolves.toBe('ok');
+  });
+
+  it('uses default parameters when none provided', async () => {
+    const fn = vi.fn().mockResolvedValue('ok');
+
+    await withRetry(fn);
+
+    expect(fn).toHaveBeenCalledTimes(1);
+  });
+});
